+++ src/js.rs
@@ -5,10 +5,10 @@ mod js_internal {
 
     #[link(wasm_import_module = "dag_viewer_js")]
     unsafe extern "C" {
-        pub fn js_fill_circ(x: i32, y: i32, radius: i32, colour: Colour);
-        pub fn js_fill_rect(x: i32, y: i32, w: i32, h: i32, colour: Colour);
-        pub fn js_fill_line(x1: i32, y1: i32, x2: i32, y2: i32, colour: Colour, width: i32);
-        pub fn js_fill_string(x: i32, y: i32, s: *const u8, colour: Colour, size: i32);
+        pub fn js_fill_circ(x: f32, y: f32, radius: f32, colour: Colour);
+        pub fn js_fill_rect(x: f32, y: f32, w: f32, h: f32, colour: Colour);
+        pub fn js_fill_line(x1: f32, y1: f32, x2: f32, y2: f32, colour: Colour, width: f32);
+        pub fn js_fill_string(x: f32, y: f32, s: *const u8, colour: Colour, size: f32);
 
         // debugging functions
         #[allow(dead_code)]
@@ -18,19 +18,19 @@ mod js_internal {
     }
 }
 
-pub fn fill_circ(x: i32, y: i32, radius: i32, colour: Colour) -> () {
+pub fn fill_circ(x: f32, y: f32, radius: f32, colour: Colour) -> () {
     unsafe { js_internal::js_fill_circ(x, y, radius, colour) }
 }
 
-pub fn fill_rect(x: i32, y: i32, w: i32, h: i32, colour: Colour) -> () {
+pub fn fill_rect(x: f32, y: f32, w: f32, h: f32, colour: Colour) -> () {
     unsafe { js_internal::js_fill_rect(x, y, w, h, colour) }
 }
 
-pub fn fill_line(x1: i32, y1: i32, x2: i32, y2: i32, colour: Colour, width: i32) -> () {
+pub fn fill_line(x1: f32, y1: f32, x2: f32, y2: f32, colour: Colour, width: f32) -> () {
     unsafe { js_internal::js_fill_line(x1, y1, x2, y2, colour, width) }
 }
 
-pub fn fill_string(x: i32, y: i32, s: *const u8, colour: Colour, size: i32) -> () {
+pub fn fill_string(x: f32, y: f32, s: *const u8, colour: Colour, size: f32) -> () {
     unsafe { js_internal::js_fill_string(x, y, s, colour, size) }
 }
